requirements.txt
flask
python-docx
lxml
gunicorn
PyMuPDF
Pillow